\name{PSgraphing-package}
\alias{PSgraphing-package}
\alias{PSgraphing}
\docType{package}
\title{
\packageTitle{PSgraphing}
}
\description{
\packageDescription{PSgraphing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PSgraphing}
\packageIndices{PSgraphing}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PSgraphing}

Maintainer: \packageMaintainer{PSgraphing}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
