% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PSgraphing.R
\name{PSgraphing}
\alias{PSgraphing}
\title{A propensity score graph}
\usage{
PSgraphing(estimate, count, text, name)
}
\arguments{
\item{pass}{a string to text to express the n range per decile, name is the graph output's name}
}
\description{
A propensity score graph
}
\examples{
data(PSgraphing)
PSgraphing(count=Count_data3, estimate=Estimate_RR, text='(total n range: 3237 to 3245 per decile)', name='myPSgraph.pdf')
}
\keyword{graph}
\keyword{propensity}
\keyword{score}

